requirements.txt
huggingface-hub == 0.0.17
numpy == 1.25
matplotlib >= 3.6.2
pyyaml >= 6.0
rasterio >= 1.3.4
scikit-image >= 0.19.3
segmentation-models-pytorch==0.3.2
six
torchmetrics == 1.1
transformers == 4.34.0